refactor: extract remaining files in UnitAndLessonLists into its own component folders
