+++ src/components/TeacherComponents/ListItemCard/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./ListItemCard";
