+++ src/components/TeacherComponents/ListItemHeader/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./ListItemHeader";
