+++ src/components/TeacherComponents/ListItemHeaderCategoryHeading/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./ListItemHeaderCategoryHeading";
