+++ src/components/TeacherComponents/ListItemHeaderExpemplarCategoryHeading/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./ListItemHeaderExpemplarCategoryHeading";
