+++ src/components/TeacherComponents/ListItemIndexDesktop/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./ListItemIndexDesktop";
