+++ src/components/TeacherComponents/UnitListItemIconDesktop/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./UnitListItemIconDesktop";
