+++ src/components/TeacherComponents/UnitListItemIconMobile/UnitListItemIconMobile.tsx
@@ -1,43 +1,40 @@
 import { FC } from "react";
 
-import SubjectIcon from "../SharedComponents/SubjectIcon";
-
+import SubjectIcon from "@/components/SharedComponents/SubjectIcon";
 import { OakColorName } from "@/styles/theme/types";
 import Flex from "@/components/SharedComponents/Flex";
 
-export type ListItemIconDesktopProps = {
-  title: string;
+export type UnitListItemIconMobileProps = {
   subjectSlug: string;
   background: OakColorName;
 };
 
-const ListItemIconDesktop: FC<ListItemIconDesktopProps> = (props) => {
+const UnitListItemIconMobile: FC<UnitListItemIconMobileProps> = (props) => {
   const { background, subjectSlug } = props;
 
   return (
     <Flex
       $justifyContent={"center"}
-      $display={["none", "flex"]}
+      $display={["flex", "none"]}
       $alignItems={"center"}
-      $minHeight={110}
-      $minWidth={130}
       $height={"100%"}
+      $minHeight={72}
+      $minWidth={72}
       $background={background}
       $position={"relative"}
-      $dropShadow={"subjectCard"}
-      $transition={"all 0.4s ease-out"}
-      data-testid="list-item-icon"
+      $ml={"auto"}
     >
       <SubjectIcon
         subjectSlug={subjectSlug}
         height={96}
         width={96}
-        $maxHeight={96}
-        $maxWidth={96}
+        $height="auto"
+        $maxHeight={72}
+        $maxWidth={72}
         $ma={"auto"}
       />
     </Flex>
   );
 };
 
-export default ListItemIconDesktop;
+export default UnitListItemIconMobile;
